requirements.txt
tweepy>=4.14.0
python-dotenv>=1.0.0
requests>=2.31.0
pillow>=10.4.0
schedule>=1.2.0
flask>=2.3.3
openai>=1.3.0
python-decouple>=3.8
google-generativeai>=0.3.0
unsplash-python>=1.1.0